Move the format date string from the controller to the framework utils
